+++ MathGame/MathGame.Models/Enums/Difficulty.cs
@@ -1,0 +1,9 @@
+﻿namespace MathGame.Models.Enums
+{
+    public enum Difficulty
+    {
+        Easy,
+        Medium,
+        Hard
+    }
+}
