+++ MathGame/MathGame.Models/Enums/GameType.cs
@@ -1,0 +1,11 @@
+﻿namespace MathGame.Models.Enums
+{
+    public enum GameType
+    {
+        Addition,
+        Subtraction,
+        Multiplication,
+        Division,
+        Mixed
+    }
+}
