+++ MathGame/MathGame.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35122.118
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame", "MathGame\MathGame.csproj", "{0936048C-66E3-447B-870E-449EFB9B0264}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathGame", "MathGame\MathGame.csproj", "{0936048C-66E3-447B-870E-449EFB9B0264}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame.Models", "MathGame.Models\MathGame.Models.csproj", "{15C89D88-E19D-46E4-88B0-4E75E2503D9D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{0936048C-66E3-447B-870E-449EFB9B0264}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0936048C-66E3-447B-870E-449EFB9B0264}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0936048C-66E3-447B-870E-449EFB9B0264}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15C89D88-E19D-46E4-88B0-4E75E2503D9D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15C89D88-E19D-46E4-88B0-4E75E2503D9D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15C89D88-E19D-46E4-88B0-4E75E2503D9D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15C89D88-E19D-46E4-88B0-4E75E2503D9D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
